requirements.txt
flask
python-pptx
deep-translator
tqdm
gunicorn
